Improve image extractor
Make it so it first tries to get the local daemon and fall back to the
remote one, so it behaves like the docker client and we alwasy first try
to get it from the local daemon.

Also add retry backoff.

Retry backoff by: Fredrik Lönnegren fredrik.lonnegren@suse.com

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>