Improve GetImage platform handling
local daemon is not able to pull multiple arch images nor you can ask it
to give you an image and a platform. So if the local daemon has image X
for platform amd64 and you wanted it for arm64 it would return that
wrong platform image.

This commit checks if the platform is the proper one and if not falls
back into asking the remote for the proper platform image

Signed-off-by: Itxaka <itxaka@kairos.io>